refactor: threshold data-type to float and removed ability to update sensorStation measurementInterval

+++ src/main/java/at/qe/skeleton/dtos/ThresholdCreateDTO.java
@@ -16,7 +16,7 @@ public record ThresholdCreateDTO(
         Metric metric,
 
         @NotNull
-        Long boundValue,
+        Float boundValue,
 
         @NotNull
         ThresholdType thresholdType,
